+++ examples/fizzbuzz.py
@@ -1,26 +1,27 @@
 # Copyright (c) 2014 Mick Krippendorf <m.krippendorf+hornet@posteo.de>
+# SPDX-License-Identifier: MIT
 
 from toolz import take
 
-from hornet import database
+from hornet import DCG, database
 from hornet.symbols import (
-    Current,
     M,
     N,
-    Next,
     R,
-    Result,
     S,
-    W,
     Ws,
+    _,
+    cut,
     equal,
     fb,
-    findall,
     fizzbuzz,
     ifelse,
+    inline,
     join,
     let,
+    phrase,
     word,
+    words,
 )
 
 db = database()
@@ -28,46 +29,32 @@ db = database()
 
 def main():
     db.tell(
-        word("fizz", N).when(
-            let(M, N % 3),
-            equal(M, 0),
-        ),
-        word("buzz", N).when(
-            let(M, N % 5),
-            equal(M, 0),
-        ),
-        word("blub", N).when(
-            let(M, N % 7),
-            equal(M, 0),
-        ),
-        fb(N, Result).when(
-            findall(W, word(W, N), Ws),
+        DCG(words(N)).when(word(3, N), word(5, N), word(7, N), inline(cut)),
+        DCG(word(3, N)).when(inline(let(M, N % 3), equal(M, 0)), ["fizz"]),
+        DCG(word(5, N)).when(inline(let(M, N % 5), equal(M, 0)), ["buzz"]),
+        DCG(word(7, N)).when(inline(let(M, N % 7), equal(M, 0)), ["quux"]),
+        DCG(word(_, _)).when(),
+        fb(N, R).when(
+            phrase(words(N), Ws),
             join(Ws, S),
             ifelse(
                 equal(S, ""),
-                equal(N, Result),
-                equal(S, Result),
+                equal(N, R),
+                equal(S, R),
             ),
         ),
-        fb(Current, Result).when(
-            let(Next, Current + 1),
-            fb(Next, Result),
+        fb(N, R).when(
+            let(M, N + 1),
+            fb(M, R),
         ),
-        fizzbuzz(Result).when(
-            fb(1, Result),
+        fizzbuzz(R).when(
+            fb(1, R),
         ),
     )
 
-    try:
-        for s in take(1111, db.ask(fizzbuzz(R))):
-            print(s[R])
-            pass
-    except RuntimeError:
-        print("oops!")
+    for s in take(111, db.ask(fizzbuzz(R))):
+        print(s[R])
 
 
 if __name__ == "__main__":
-    # from examples.utils import timer
-    #
-    # with timer():
     main()
