+++ src/hornet/__init__.py
@@ -5,20 +5,18 @@ from __future__ import annotations
 
 from . import combinators, expressions, symbols, terms
 from ._builtins import _bootstrap_database
-from .combinators import Emit, Goal, Next, Result, Step, Subst
+from .clauses import predicate
 from .expressions import DCG
-from .goals import Database, predicate
 from .symbols import (
     append,
     arithmetic_equal,
-    arithmetic_not_equal,
-    atomic,
     call,
     cut,
     equal,
     fail,
     findall,
     greater,
+    ifelse,
     ignore,
     is_atom,
     is_atomic,
@@ -34,22 +32,18 @@ from .symbols import (
     join,
     length,
     let,
-    listing,
     lwriteln,
     maplist,
     member,
     nl,
     nonvar,
-    numeric,
     once,
-    real,
+    phrase,
     repeat,
     reverse,
     select,
     smaller,
-    tail,
     throw,
-    transpose,
     true,
     unequal,
     univ,
@@ -58,36 +52,28 @@ from .symbols import (
 )
 
 __all__ = (
-    "DCG",
-    "database",
-    "is_constant",
     "combinators",
     "expressions",
     "symbols",
     "terms",
-    "Database",
-    "Emit",
-    "Goal",
-    "Next",
-    "Result",
-    "Step",
-    "Subst",
+    "DCG",
+    "database",
     "append",
     "arithmetic_equal",
-    "arithmetic_not_equal",
-    "atomic",
     "call",
     "cut",
     "equal",
     "fail",
     "findall",
     "greater",
+    "ifelse",
     "ignore",
     "is_atom",
     "is_atomic",
     "is_bool",
     "is_bytes",
     "is_complex",
+    "is_constant",
     "is_float",
     "is_int",
     "is_numeric",
@@ -96,23 +82,19 @@ __all__ = (
     "join",
     "length",
     "let",
-    "listing",
     "lwriteln",
     "maplist",
     "member",
     "nl",
     "nonvar",
-    "numeric",
     "once",
+    "phrase",
     "predicate",
-    "real",
     "repeat",
     "reverse",
     "select",
     "smaller",
-    "tail",
     "throw",
-    "transpose",
     "true",
     "unequal",
     "univ",
