+++ src/hornet/expressions.py
@@ -6,11 +6,11 @@ from __future__ import annotations
 from dataclasses import dataclass
 from functools import partial, reduce
 from itertools import count
-from typing import Any, Callable, Iterator, Protocol, cast
+from typing import Any, Callable, Iterator, cast
 
-from toolz.functoolz import compose, flip
+from toolz.functoolz import compose
 
-from .states import StateGenerator, const, get_state, identity, set_state, with_state
+from .states import StateGenerator, const, get_state, set_state, with_state
 from .terms import (
     EMPTY,
     Add,
@@ -44,27 +44,204 @@ from .terms import (
 )
 
 
-class HasTerm[T](Protocol):
-    @property
-    def term(self) -> T: ...
+@dataclass(frozen=True, slots=True)
+class RuleExpression:
+    term: Term
+
+
+@dataclass(frozen=True, slots=True)
+class DCG:
+    expr: Expression[Atom | Functor]
+
+    def when(self, *args) -> RuleExpression:
+        head = self.expr.term
+        body = Conjunction(*(promote(arg) for arg in args))
+        new_head, new_body = dcg_expand(head, body)
+        assert isinstance(new_head, MatchTerm)
+        assert isinstance(new_body, QueryTerm)
+        return RuleExpression(term=HornetRule(head=new_head, body=new_body))
+
+
+@dataclass(frozen=True, slots=True)
+class Expression[T: Term]:
+    """
+    An Expression object is a monadic wrapper around a Term.
+    """
+
+    term: T
+
+    def __init__(self, term: T):
+        object.__setattr__(self, "term", promote(term))
+
+    def when(self, *args) -> RuleExpression:
+        head = self.term
+        body = Conjunction(*(promote(arg) for arg in args))
+        assert isinstance(head, MatchTerm)
+        assert isinstance(body, QueryTerm)
+        return RuleExpression(HornetRule(head=head, body=body))
+
+    def __repr__(self):
+        return f"Expression({repr(self.term)})"
+
+    def __str__(self):
+        return repr(self.term)
+
+    def __eq__(self, other):
+        return isinstance(other, Expression) and self.term == other.term
+
+    def __hash__(self) -> int:
+        return hash((Expression, self.term))
+
+    def __neg__(self):
+        return expression(USub(promote(self)))
+
+    def __pos__(self):
+        return expression(UAdd(promote(self)))
+
+    def __invert__(self):
+        return expression(Invert(promote(self)))
+
+    def __add__(self, other):
+        return expression(Add(promote(self), promote(other)))
+
+    def __radd__(self, other):
+        return expression(Add(promote(other), promote(self)))
+
+    def __sub__(self, other):
+        return expression(Sub(promote(self), promote(other)))
+
+    def __rsub__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __mul__(self, other):
+        return expression(Mul(promote(self), promote(other)))
+
+    def __rmul__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __truediv__(self, other):
+        return expression(Div(promote(self), promote(other)))
+
+    def __rtruediv__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __floordiv__(self, other):
+        return expression(FloorDiv(promote(self), promote(other)))
+
+    def __rfloordiv__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __mod__(self, other):
+        return expression(Mod(promote(self), promote(other)))
+
+    def __rmod__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __pow__(self, other):
+        return expression(Pow(promote(self), promote(other)))
+
+    def __rpow__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __lshift__(self, other):
+        return expression(LShift(promote(self), promote(other)))
+
+    def __rlshift__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __rshift__(self, other):
+        return expression(RShift(promote(self), promote(other)))
+
+    def __rrshift__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __and__(self, other):
+        return expression(BitAnd(promote(self), promote(other)))
+
+    def __rand__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __xor__(self, other):
+        return expression(BitXor(promote(self), promote(other)))
+
+    def __rxor__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __or__(self, other):
+        return expression(BitOr(promote(self), promote(other)))
+
+    def __ror__(self, other):
+        return expression(Sub(promote(other), promote(self)))
+
+    def __call__(self, *args):
+        match promote(self):
+            case Atom(name):
+                return expression((Functor(name, *(promote(arg) for arg in args))))
+
+            case _:
+                raise TypeError(f"Atom required, not {self}")
+
+
+def promote(obj: Any) -> Term | tuple:
+    """
+    Convert a Python object to a Term.
+    """
+    match obj:
+        case (
+            Wildcard()
+            | Variable()
+            | Atom()
+            | Empty()
+            | str()
+            | bytes()
+            | int()
+            | bool()
+            | float()
+            | complex()
+        ):
+            return obj
+        case Functor(name=name, args=args):
+            return Functor(name, *(promote(arg) for arg in args))
+        case Structure(args):
+            return type(obj)(*(promote(arg) for arg in args))
+        case Expression(term):
+            return promote(term)
+        case tuple():
+            return tuple(promote(each) for each in obj)
+        case Cons(head=BitOr(left=left, right=right)):
+            return Cons(promote(left), promote(right))
+        case []:
+            return EMPTY
+        case [Expression(BitOr(left=left, right=right)), *tail]:
+            assert not tail
+            return Cons(promote(left), promote(right))
+        case [head, *tail]:
+            return Cons(promote(head), promote(list(tail)))
+        case _:
+            raise TypeError(f"{type(obj)}")
 
 
 type VarCount = tuple[int, Iterator[int]]
 _v_count: Iterator[int] = count()
 
 
+def dcg_expand(head: Term, body: Term) -> tuple[Term, Term]:
+    (new_head, new_body), _ = _dcg_expand(head, body).run((next(_v_count), _v_count))
+    return new_head, new_body
+
+
 @with_state
 def _dcg_expand(head: Term, body: Term) -> StateGenerator[VarCount, tuple[Term, Term]]:
     @with_state
     def current_variable() -> StateGenerator[VarCount, Variable]:
-        i, _ = yield get_state(identity)
+        i, _ = yield get_state()
         return Variable(f"S${i}")
 
     @with_state
     def advance_variables() -> StateGenerator[VarCount, tuple[Variable, Variable]]:
-        i, c = yield get_state(identity)
-        j = next(c)
-        yield set_state(const((j, c)))
+        i, counter = yield get_state()
+        j = next(counter)
+        yield set_state(const((j, counter)))
         return Variable(f"S${i}"), Variable(f"S${j}")
 
     @with_state
@@ -135,188 +312,6 @@ def _dcg_expand(head: Term, body: Term) -> StateGenerator[VarCount, tuple[Term, 
     assert isinstance(head_expanded, MatchTerm)
     assert isinstance(body_expanded, QueryTerm)
     return head_expanded, body_expanded
-
-
-def dcg_expand(head: Term, body: Term) -> tuple[Term, Term]:
-    (new_head, new_body), _ = _dcg_expand(head, body).run((next(_v_count), _v_count))
-    return new_head, new_body
-
-
-@dataclass(frozen=True, slots=True)
-class RuleExpression:
-    term: Term
-
-
-@dataclass(frozen=True, slots=True)
-class DCG:
-    expr: Expression[Atom | Functor]
-
-    def when(self, *args) -> RuleExpression:
-        head = self.expr.term
-        body = Conjunction(*(promote(arg) for arg in args))
-        new_head, new_body = dcg_expand(head, body)
-        assert isinstance(new_head, MatchTerm)
-        assert isinstance(new_body, QueryTerm)
-        return RuleExpression(term=HornetRule(head=new_head, body=new_body))
-
-
-@dataclass(frozen=True, slots=True)
-class Expression[T: Term]:
-    """
-    An Expression object is a monadic wrapper around a Term.
-    """
-
-    term: T
-
-    def when(self, *args) -> RuleExpression:
-        head = self.term
-        body = Conjunction(*(promote(arg) for arg in args))
-        assert isinstance(head, MatchTerm)
-        assert isinstance(body, QueryTerm)
-        return RuleExpression(HornetRule(head=head, body=body))
-
-    def __repr__(self):
-        return f"Expression({repr(self.term)})"
-
-    def __str__(self):
-        return f"Expression({str(self.term)})"
-
-    def __eq__(self, other):
-        return isinstance(other, Expression) and self.term == other.term
-
-    def __hash__(self) -> int:
-        return hash((Expression, self.term))
-
-    def __neg__(self):
-        return expression(USub(promote(self)))
-
-    def __pos__(self):
-        return expression(UAdd(promote(self)))
-
-    def __invert__(self):
-        return expression(Invert(promote(self)))
-
-    def __add__(self, other):
-        return expression(Add(promote(self), promote(other)))
-
-    @flip
-    def __radd__(self, other):
-        return expression(Add(promote(other), promote(self)))
-
-    def __sub__(self, other):
-        return expression(Sub(promote(self), promote(other)))
-
-    @flip
-    def __rsub__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __mul__(self, other):
-        return expression(Mul(promote(self), promote(other)))
-
-    @flip
-    def __rmul__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __truediv__(self, other):
-        return expression(Div(promote(self), promote(other)))
-
-    @flip
-    def __rtruediv__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __floordiv__(self, other):
-        return expression(FloorDiv(promote(self), promote(other)))
-
-    @flip
-    def __rfloordiv__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __mod__(self, other):
-        return expression(Mod(promote(self), promote(other)))
-
-    @flip
-    def __rmod__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __pow__(self, other):
-        return expression(Pow(promote(self), promote(other)))
-
-    @flip
-    def __rpow__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __lshift__(self, other):
-        return expression(LShift(promote(self), promote(other)))
-
-    @flip
-    def __rlshift__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __rshift__(self, other):
-        return expression(RShift(promote(self), promote(other)))
-
-    @flip
-    def __rrshift__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __and__(self, other):
-        return expression(BitAnd(promote(self), promote(other)))
-
-    @flip
-    def __rand__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __xor__(self, other):
-        return expression(BitXor(promote(self), promote(other)))
-
-    @flip
-    def __rxor__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __or__(self, other):
-        return expression(BitOr(promote(self), promote(other)))
-
-    @flip
-    def __ror__(self, other):
-        return expression(Sub(promote(other), promote(self)))
-
-    def __call__(self, *args):
-        match promote(self):
-            case Atom(name):
-                return expression((Functor(name, *(promote(arg) for arg in args))))
-
-            case _:
-                raise TypeError(f"Atom required, not {self}")
-
-
-def promote(obj: Any) -> Term | tuple:
-    """
-    Convert a Python object to a Term.
-    """
-    match obj:
-        case (
-            Wildcard()
-            | Variable()
-            | Atom()
-            | Structure()
-            | str()
-            | bytes()
-            | int()
-            | bool()
-            | float()
-            | complex()
-        ):
-            return obj
-        case Expression(term):
-            return term
-        case tuple():
-            return tuple(promote(each) for each in obj)
-        case list([]):
-            return EMPTY
-        case list([head, *tail]):
-            return Cons(promote(head), promote(list(tail)))
-        case _:
-            raise TypeError(f"{type(obj)}")
 
 
 # In the Monad, unit is the same as Expression:
